+++ packages/tendermint-rpc/src/adaptors/v0-33/hasher.spec.ts
@@ -2,7 +2,6 @@ import { fromBase64, fromHex } from "@cosmjs/encoding";
 import { ReadonlyDate } from "readonly-date";
 
 import { ReadonlyDateWithNanoseconds } from "../../responses";
-import { TxBytes } from "../../types";
 import { hashBlock, hashTx } from "./hasher";
 
 describe("Hasher", () => {
@@ -10,7 +9,7 @@ describe("Hasher", () => {
     // This was taken from a result from /tx_search of some random test transaction
     // curl "http://localhost:11127/tx_search?query=\"tx.hash='5CB2CF94A1097A4BC19258BC2353C3E76102B6D528458BE45C855DC5563C1DB2'\""
     const txId = fromHex("5CB2CF94A1097A4BC19258BC2353C3E76102B6D528458BE45C855DC5563C1DB2");
-    const txData = fromBase64("YUpxZDY2NURaUDMxPWd2TzBPdnNrVWFWYg==") as TxBytes;
+    const txData = fromBase64("YUpxZDY2NURaUDMxPWd2TzBPdnNrVWFWYg==");
     expect(hashTx(txData)).toEqual(txId);
   });