+++ packages/tendermint-rpc/src/adaptors/v0-33/responses.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import { fromHex } from "@cosmjs/encoding";
+import { fromBase64, fromHex } from "@cosmjs/encoding";
 import { JsonRpcSuccessResponse } from "@cosmjs/json-rpc";
 
 import {
@@ -10,20 +10,15 @@ import {
   assertObject,
   assertSet,
   assertString,
-  Base64,
-  Base64String,
   DateTime,
   dictionaryToStringMap,
-  Hex,
-  HexString,
   Integer,
-  IntegerString,
   may,
   optional,
 } from "../../encodings";
 import * as responses from "../../responses";
 import { SubscriptionEvent } from "../../rpcclients";
-import { IpPortString, TxBytes, TxHash, ValidatorPubkey, ValidatorSignature } from "../../types";
+import { ValidatorPubkey, ValidatorSignature } from "../../types";
 import { hashTx } from "./hasher";
 
 interface AbciInfoResult {
@@ -32,15 +27,16 @@ interface AbciInfoResult {
 
 interface RpcAbciInfoResponse {
   readonly data?: string;
-  readonly last_block_height?: IntegerString;
-  readonly last_block_app_hash?: Base64String;
+  readonly last_block_height?: string;
+  /** base64 encoded */
+  readonly last_block_app_hash?: string;
 }
 
 function decodeAbciInfo(data: RpcAbciInfoResponse): responses.AbciInfoResponse {
   return {
     data: data.data,
     lastBlockHeight: may(Integer.parse, data.last_block_height),
-    lastBlockAppHash: may(Base64.decode, data.last_block_app_hash),
+    lastBlockAppHash: may(fromBase64, data.last_block_app_hash),
   };
 }
 
@@ -50,8 +46,10 @@ interface AbciQueryResult {
 
 export interface RpcProofOp {
   readonly type: string;
-  readonly key: Base64String;
-  readonly data: Base64String;
+  /** base64 encoded */
+  readonly key: string;
+  /** base64 encoded */
+  readonly data: string;
 }
 
 export interface RpcQueryProof {
@@ -62,26 +60,28 @@ function decodeQueryProof(data: RpcQueryProof): responses.QueryProof {
   return {
     ops: data.ops.map((op) => ({
       type: op.type,
-      key: Base64.decode(op.key),
-      data: Base64.decode(op.data),
+      key: fromBase64(op.key),
+      data: fromBase64(op.data),
     })),
   };
 }
 
 interface RpcAbciQueryResponse {
-  readonly key: Base64String;
-  readonly value?: Base64String;
+  /** base64 encoded */
+  readonly key: string;
+  /** base64 encoded */
+  readonly value?: string;
   readonly proof?: RpcQueryProof;
-  readonly height?: IntegerString;
-  readonly index?: IntegerString;
-  readonly code?: IntegerString; // only for errors
+  readonly height?: string;
+  readonly index?: string;
+  readonly code?: string; // only for errors
   readonly log?: string;
 }
 
 function decodeAbciQuery(data: RpcAbciQueryResponse): responses.AbciQueryResponse {
   return {
-    key: Base64.decode(optional(data.key, "" as Base64String)),
-    value: Base64.decode(optional(data.value, "" as Base64String)),
+    key: fromBase64(optional(data.key, "")),
+    value: fromBase64(optional(data.value, "")),
     proof: may(decodeQueryProof, data.proof),
     height: may(Integer.parse, data.height),
     code: may(Integer.parse, data.code),
@@ -91,14 +91,16 @@ function decodeAbciQuery(data: RpcAbciQueryResponse): responses.AbciQueryRespons
 }
 
 interface RpcAttribute {
-  readonly key: Base64String;
-  readonly value: Base64String;
+  /** base64 encoded */
+  readonly key: string;
+  /** base64 encoded */
+  readonly value: string;
 }
 
 function decodeAttribute(attribute: RpcAttribute): responses.Attribute {
   return {
-    key: Base64.decode(assertNotEmpty(attribute.key)),
-    value: Base64.decode(assertNotEmpty(attribute.value)),
+    key: fromBase64(assertNotEmpty(attribute.key)),
+    value: fromBase64(assertNotEmpty(attribute.value)),
   };
 }
 
@@ -125,13 +127,14 @@ function decodeEvents(events: readonly RpcEvent[]): readonly responses.Event[] {
 interface RpcTxData {
   readonly code?: number;
   readonly log?: string;
-  readonly data?: Base64String;
+  /** base64 encoded */
+  readonly data?: string;
   readonly events: readonly RpcEvent[];
 }
 
 function decodeTxData(data: RpcTxData): responses.TxData {
   return {
-    data: may(Base64.decode, data.data),
+    data: may(fromBase64, data.data),
     log: data.log,
     code: Integer.parse(assertNumber(optional<number>(data.code, 0))),
     events: decodeEvents(data.events),
@@ -141,7 +144,8 @@ function decodeTxData(data: RpcTxData): responses.TxData {
 // yes, a different format for status and dump consensus state
 interface RpcPubkey {
   readonly type: string;
-  readonly value: Base64String;
+  /** base64 encoded */
+  readonly value: string;
 }
 
 function decodePubkey(data: RpcPubkey): ValidatorPubkey {
@@ -149,7 +153,7 @@ function decodePubkey(data: RpcPubkey): ValidatorPubkey {
     // go-amino special code
     return {
       algorithm: "ed25519",
-      data: Base64.decode(assertNotEmpty(data.value)),
+      data: fromBase64(assertNotEmpty(data.value)),
     };
   }
   throw new Error(`unknown pubkey type: ${data.type}`);
@@ -157,22 +161,23 @@ function decodePubkey(data: RpcPubkey): ValidatorPubkey {
 
 // for evidence, block results, etc.
 interface RpcValidatorUpdate {
-  readonly address: HexString;
+  /** hex encoded */
+  readonly address: string;
   readonly pub_key: RpcPubkey;
-  readonly voting_power: IntegerString;
+  readonly voting_power: string;
 }
 
 function decodeValidatorUpdate(data: RpcValidatorUpdate): responses.Validator {
   return {
     pubkey: decodePubkey(assertObject(data.pub_key)),
     votingPower: Integer.parse(assertNotEmpty(data.voting_power)),
-    address: Hex.decode(assertNotEmpty(data.address)),
+    address: fromHex(assertNotEmpty(data.address)),
   };
 }
 
 interface RpcBlockParams {
-  readonly max_bytes: IntegerString;
-  readonly max_gas: IntegerString;
+  readonly max_bytes: string;
+  readonly max_gas: string;
 }
 
 /**
@@ -189,8 +194,8 @@ function decodeBlockParams(data: RpcBlockParams): responses.BlockParams {
 }
 
 interface RpcEvidenceParams {
-  readonly max_age_num_blocks: IntegerString;
-  readonly max_age_duration: IntegerString;
+  readonly max_age_num_blocks: string;
+  readonly max_age_duration: string;
 }
 
 function decodeEvidenceParams(data: RpcEvidenceParams): responses.EvidenceParams {
@@ -232,7 +237,7 @@ function decodeConsensusParams(data: RpcConsensusParams): responses.ConsensusPar
 }
 
 interface RpcBlockResultsResponse {
-  readonly height: IntegerString;
+  readonly height: string;
   readonly txs_results: readonly RpcTxData[] | null;
   readonly begin_block_events: readonly RpcEvent[] | null;
   readonly end_block_events: readonly RpcEvent[] | null;
@@ -252,10 +257,12 @@ function decodeBlockResults(data: RpcBlockResultsResponse): responses.BlockResul
 }
 
 interface RpcBlockId {
-  readonly hash: HexString;
+  /** hex encoded */
+  readonly hash: string;
   readonly parts: {
-    readonly total: IntegerString;
-    readonly hash: HexString;
+    readonly total: string;
+    /** hex encoded */
+    readonly hash: string;
   };
 }
 
@@ -270,8 +277,8 @@ function decodeBlockId(data: RpcBlockId): responses.BlockId {
 }
 
 interface RpcBlockVersion {
-  readonly block: IntegerString;
-  readonly app?: IntegerString;
+  readonly block: string;
+  readonly app?: string;
 }
 
 function decodeBlockVersion(data: RpcBlockVersion): responses.Version {
@@ -284,24 +291,33 @@ function decodeBlockVersion(data: RpcBlockVersion): responses.Version {
 interface RpcHeader {
   readonly version: RpcBlockVersion;
   readonly chain_id: string;
-  readonly height: IntegerString;
+  readonly height: string;
   readonly time: string;
-  readonly num_txs: IntegerString;
-  readonly total_txs: IntegerString;
+  readonly num_txs: string;
+  readonly total_txs: string;
 
   readonly last_block_id: RpcBlockId;
 
-  readonly last_commit_hash: HexString;
-  readonly data_hash: HexString;
+  /** hex encoded */
+  readonly last_commit_hash: string;
+  /** hex encoded */
+  readonly data_hash: string;
 
-  readonly validators_hash: HexString;
-  readonly next_validators_hash: HexString;
-  readonly consensus_hash: HexString;
-  readonly app_hash: HexString;
-  readonly last_results_hash: HexString;
+  /** hex encoded */
+  readonly validators_hash: string;
+  /** hex encoded */
+  readonly next_validators_hash: string;
+  /** hex encoded */
+  readonly consensus_hash: string;
+  /** hex encoded */
+  readonly app_hash: string;
+  /** hex encoded */
+  readonly last_results_hash: string;
 
-  readonly evidence_hash: HexString;
-  readonly proposer_address: HexString;
+  /** hex encoded */
+  readonly evidence_hash: string;
+  /** hex encoded */
+  readonly proposer_address: string;
 }
 
 function decodeHeader(data: RpcHeader): responses.Header {
@@ -340,7 +356,7 @@ function decodeBlockMeta(data: RpcBlockMeta): responses.BlockMeta {
 }
 
 interface RpcBlockchainResponse {
-  readonly last_height: IntegerString;
+  readonly last_height: string;
   readonly block_metas: readonly RpcBlockMeta[];
 }
 
@@ -352,19 +368,21 @@ function decodeBlockchain(data: RpcBlockchainResponse): responses.BlockchainResp
 }
 
 interface RpcBroadcastTxSyncResponse extends RpcTxData {
-  readonly hash: HexString;
+  /** hex encoded */
+  readonly hash: string;
 }
 
 function decodeBroadcastTxSync(data: RpcBroadcastTxSyncResponse): responses.BroadcastTxSyncResponse {
   return {
     ...decodeTxData(data),
-    hash: fromHex(assertNotEmpty(data.hash)) as TxHash,
+    hash: fromHex(assertNotEmpty(data.hash)),
   };
 }
 
 interface RpcBroadcastTxCommitResponse {
-  readonly height: IntegerString;
-  readonly hash: HexString;
+  readonly height: string;
+  /** hex encoded */
+  readonly hash: string;
   readonly check_tx: RpcTxData;
   readonly deliver_tx?: RpcTxData;
 }
@@ -372,7 +390,7 @@ interface RpcBroadcastTxCommitResponse {
 function decodeBroadcastTxCommit(data: RpcBroadcastTxCommitResponse): responses.BroadcastTxCommitResponse {
   return {
     height: Integer.parse(data.height),
-    hash: fromHex(assertNotEmpty(data.hash)) as TxHash,
+    hash: fromHex(assertNotEmpty(data.hash)),
     checkTx: decodeTxData(assertObject(data.check_tx)),
     deliverTx: may(decodeTxData, data.deliver_tx),
   };
@@ -380,15 +398,17 @@ function decodeBroadcastTxCommit(data: RpcBroadcastTxCommitResponse): responses.
 
 type RpcSignature = {
   readonly block_id_flag: number;
-  readonly validator_address: HexString;
+  /** hex encoded */
+  readonly validator_address: string;
   readonly timestamp: string;
-  readonly signature: Base64String;
+  /** bae64 encoded */
+  readonly signature: string;
 };
 
 function decodeSignature(data: RpcSignature): ValidatorSignature {
   return {
     algorithm: "ed25519",
-    data: Base64.decode(assertNotEmpty(data.signature)),
+    data: fromBase64(assertNotEmpty(data.signature)),
   };
 }
 
@@ -422,7 +442,7 @@ function decodeCommitResponse(data: RpcCommitResponse): responses.CommitResponse
 
 interface RpcValidatorGenesis {
   readonly pub_key: RpcPubkey;
-  readonly power: IntegerString;
+  readonly power: string;
   readonly name?: string;
 }
 
@@ -441,7 +461,8 @@ interface RpcGenesisResponse {
   // The validators key is used to specify a set of validators for testnets or PoA blockchains.
   // PoS blockchains use the app_state.genutil.gentxs field to stake and bond a number of validators in the first block.
   readonly validators?: readonly RpcValidatorGenesis[];
-  readonly app_hash: HexString;
+  /** hex encoded */
+  readonly app_hash: string;
   readonly app_state: Record<string, unknown> | undefined;
 }
 
@@ -462,9 +483,10 @@ function decodeGenesis(data: RpcGenesisResponse): responses.GenesisResponse {
 
 // this is in status
 interface RpcValidatorInfo {
-  readonly address: HexString;
+  /** hex encoded */
+  readonly address: string;
   readonly pub_key: RpcPubkey;
-  readonly voting_power: IntegerString;
+  readonly voting_power: string;
 }
 
 function decodeValidatorInfo(data: RpcValidatorInfo): responses.Validator {
@@ -476,16 +498,18 @@ function decodeValidatorInfo(data: RpcValidatorInfo): responses.Validator {
 }
 
 interface RpcNodeInfo {
-  readonly id: HexString;
-  readonly listen_addr: IpPortString;
+  /** hex encoded */
+  readonly id: string;
+  /** IP and port */
+  readonly listen_addr: string;
   readonly network: string;
   readonly version: string;
   readonly channels: string; // ???
   readonly moniker: string;
   readonly protocol_version: {
-    readonly p2p: IntegerString;
-    readonly block: IntegerString;
-    readonly app: IntegerString;
+    readonly p2p: string;
+    readonly block: string;
+    readonly app: string;
   };
   /**
    * Additional information. E.g.
@@ -515,9 +539,11 @@ function decodeNodeInfo(data: RpcNodeInfo): responses.NodeInfo {
 }
 
 interface RpcSyncInfo {
-  readonly latest_block_hash: HexString;
-  readonly latest_app_hash: HexString;
-  readonly latest_block_height: IntegerString;
+  /** hex encoded */
+  readonly latest_block_hash: string;
+  /** hex encoded */
+  readonly latest_app_hash: string;
+  readonly latest_block_height: string;
   readonly latest_block_time: string;
   readonly catching_up: boolean;
 }
@@ -560,52 +586,58 @@ function decodeStatus(data: RpcStatusResponse): responses.StatusResponse {
  * }
  */
 interface RpcTxProof {
-  readonly data: Base64String;
-  readonly root_hash: HexString;
+  /** base64 encoded */
+  readonly data: string;
+  /** hex encoded */
+  readonly root_hash: string;
   readonly proof: {
-    readonly total: IntegerString;
-    readonly index: IntegerString;
-    readonly leaf_hash: Base64String;
-    readonly aunts: readonly Base64String[];
+    readonly total: string;
+    readonly index: string;
+    /** base64 encoded */
+    readonly leaf_hash: string;
+    /** base64 encoded */
+    readonly aunts: readonly string[];
   };
 }
 
 function decodeTxProof(data: RpcTxProof): responses.TxProof {
   return {
-    data: Base64.decode(assertNotEmpty(data.data)),
+    data: fromBase64(assertNotEmpty(data.data)),
     rootHash: fromHex(assertNotEmpty(data.root_hash)),
     proof: {
       total: Integer.parse(assertNotEmpty(data.proof.total)),
       index: Integer.parse(assertNotEmpty(data.proof.index)),
-      leafHash: Base64.decode(assertNotEmpty(data.proof.leaf_hash)),
-      aunts: assertArray(data.proof.aunts).map(Base64.decode),
+      leafHash: fromBase64(assertNotEmpty(data.proof.leaf_hash)),
+      aunts: assertArray(data.proof.aunts).map(fromBase64),
     },
   };
 }
 
 interface RpcTxResponse {
-  readonly tx: Base64String;
+  /** Raw tx bytes, base64 encoded */
+  readonly tx: string;
   readonly tx_result: RpcTxData;
-  readonly height: IntegerString;
+  readonly height: string;
   readonly index: number;
-  readonly hash: HexString;
+  /** hex encoded */
+  readonly hash: string;
   readonly proof?: RpcTxProof;
 }
 
 function decodeTxResponse(data: RpcTxResponse): responses.TxResponse {
   return {
-    tx: Base64.decode(assertNotEmpty(data.tx)) as TxBytes,
+    tx: fromBase64(assertNotEmpty(data.tx)),
     result: decodeTxData(assertObject(data.tx_result)),
     height: Integer.parse(assertNotEmpty(data.height)),
     index: Integer.parse(assertNumber(data.index)),
-    hash: fromHex(assertNotEmpty(data.hash)) as TxHash,
+    hash: fromHex(assertNotEmpty(data.hash)),
     proof: may(decodeTxProof, data.proof),
   };
 }
 
 interface RpcTxSearchResponse {
   readonly txs: readonly RpcTxResponse[];
-  readonly total_count: IntegerString;
+  readonly total_count: string;
 }
 
 function decodeTxSearch(data: RpcTxSearchResponse): responses.TxSearchResponse {
@@ -616,15 +648,16 @@ function decodeTxSearch(data: RpcTxSearchResponse): responses.TxSearchResponse {
 }
 
 interface RpcTxEvent {
-  readonly tx: Base64String;
+  /** Raw tx bytes, base64 encoded */
+  readonly tx: string;
   readonly result: RpcTxData;
-  readonly height: IntegerString;
+  readonly height: string;
   /** Not set since Tendermint 0.34 */
   readonly index?: number;
 }
 
 function decodeTxEvent(data: RpcTxEvent): responses.TxEvent {
-  const tx = Base64.decode(assertNotEmpty(data.tx)) as TxBytes;
+  const tx = fromBase64(assertNotEmpty(data.tx));
   return {
     tx: tx,
     hash: hashTx(tx),
@@ -636,7 +669,7 @@ function decodeTxEvent(data: RpcTxEvent): responses.TxEvent {
 
 // for validators
 interface RpcValidatorData extends RpcValidatorUpdate {
-  readonly accum?: IntegerString;
+  readonly accum?: string;
 }
 
 function decodeValidatorData(data: RpcValidatorData): responses.Validator {
@@ -647,7 +680,7 @@ function decodeValidatorData(data: RpcValidatorData): responses.Validator {
 }
 
 interface RpcValidatorsResponse {
-  readonly block_height: IntegerString;
+  readonly block_height: string;
   readonly validators: readonly RpcValidatorData[];
 }
 
@@ -661,9 +694,9 @@ function decodeValidators(data: RpcValidatorsResponse): responses.ValidatorsResp
 interface RpcEvidence {
   readonly type: string;
   readonly validator: RpcValidatorUpdate;
-  readonly height: IntegerString;
-  readonly time: IntegerString;
-  readonly totalVotingPower: IntegerString;
+  readonly height: string;
+  readonly time: string;
+  readonly totalVotingPower: string;
 }
 
 function decodeEvidence(data: RpcEvidence): responses.Evidence {
@@ -684,7 +717,8 @@ interface RpcBlock {
   readonly header: RpcHeader;
   readonly last_commit: RpcCommit;
   readonly data: {
-    readonly txs?: readonly Base64String[];
+    /** Raw tx bytes, base64 encoded */
+    readonly txs?: readonly string[];
   };
   readonly evidence?: {
     readonly evidence?: readonly RpcEvidence[];
@@ -695,7 +729,7 @@ function decodeBlock(data: RpcBlock): responses.Block {
   return {
     header: decodeHeader(assertObject(data.header)),
     lastCommit: decodeCommit(assertObject(data.last_commit)),
-    txs: data.data.txs ? assertArray(data.data.txs).map(Base64.decode) : [],
+    txs: data.data.txs ? assertArray(data.data.txs).map(fromBase64) : [],
     evidence: data.evidence && may(decodeEvidences, data.evidence.evidence),
   };
 }